Refactor feature selection logic: remove old implementation and add new structure for better organization and performance
